init commit

+++ src/NotOrtalamasi.java
@@ -19,7 +19,7 @@ public class NotOrtalamasi {
         System.out.print("muzik notunuz : ");
         muzik = scanner.nextInt();
          toplam = matematik+fizik+kimya+turkce+tarih+muzik;
-         ortalama = (double) (toplam/6.0);
+         ortalama =(toplam/6.0);
         System.out.println("not ortalamasi : "+ ortalama);
         if (ortalama>60){
             System.out.println("Sinifi gecti");
